Factored room header buttons into separate files
